Access database._security document
Currently, trying to read or write `_security` document using
`db["_security"]` fails, because aiocouch tries to access it as a
regular document. But `_security` is special: it doesn't have `_id` not
`_rev` (it's one of the only documents without revisions).

I propose to implement wrappers to get and set this document:

    async with CouchDB() as couch:
        db = await couch["db"]
        print(await db.get_security())
        await db.set_security({
            "members": {"names": ["user1"],
                        "roles": ["db_reader", "db_writer"]},
        })