Implement Document.fetch_info() to issue a HEAD request
There are situations when it's useful to send a tiny HTTP `HEAD /db/doc`
request, to retrieve only the document current revision.

I propose to add a method to retrieve this info:

    >>> print(await doc.fetch_info())
    {"ok": true,
     "id": "SpaghettiWithMeatballs",
     "rev": "1-917fa2381192822767f010b95b45325b"}

The output is similar to what a HTTP `PUT /db/doc` would return.

Note: the new `return_response` argument for `RemoteServer._request()`
will be reused in a future commit (for attachments).