Fix crash on Document.save() if conflict but not _rev
A common pattern is to try to save a document, and pass if it exists:

    try:
        await Document(db, "id").save()
    except aiocouch.exception.ConflictError:
        pass

Currently, in case of conflict, aiocouch fails with:

    [...]
    aiohttp.client_exceptions.ClientResponseError: 409...
    During handling of the above exception, another exception occurred:
    [...]
    KeyError: 'rev'

I propose to set `rev = None` in error handling, if not available.
Another solution would be to always define `Document._data["_rev"]`, but
this has other implications.