feat bulk operation: don't crash on pristine document
`Database.update_docs(ids=[...])` operation was working without crash only if
all documents were either modified or created.
But if only a subset of documents are really modified inside the bulk operation
it crashed with the error:
```
File "/var/coatl/lib64/python3.8/site-packages/aiocouch/bulk.py", line 59, in __aexit__
	assert status["id"] == doc.id
AssertionError
```
I think it's cool to improve it because:
If you have `foo` and `baz` documents in your db, and want to update both,
but `foo` already have the update. It's a nice shortcut to update both
without having to exclude `foo` from the bulk operation (see unit test for
an example).