feat: Add revs() method to fetch document known revs
This commit adds a way to list known revisions for a document on the server.
Known revisions are fetched on the server by making a GET request of the
document with `revs_info=true` in the request parameters.

note: we could also use `revs=true` in the request to get similar results, but
      the CouchDB response in this case is harder to process. The result is the
      same in both case, so I went with the option leading to less code.

See the [CouchDB documentation for fetching documents] for more info on this.

[CouchDB documentation for fetching documents]:
https://docs.couchdb.org/en/stable/api/document/common.html#get--db-docid

Initially discussed in #58