Changes handling of key and keys parameter in View
The `key` and `keys` parameter won't be accepted by the high-level
methods `ids()` and `docs()` of the `View` class. Use the `ids`
parameter instead. If you insist on using them, use the `get()` method
instead. Make sure to pre-format the values, as CouchDB expects these
to be strings containing JSON.

Fixes #57.